Added 'WaterLogManager' class
ViewController 내에 있던 WaterLog 관련 로직을 WaterLogManager클래스를
만들어서 구현하였다. 사용하던 메서드들은 클래스 메서드로써 사용이 된다.
