+++ waterful/ViewController.swift
@@ -47,36 +47,41 @@ class ViewController: UIViewController, WCSessionDelegate {
     @IBOutlet weak var bottleLabel: UILabel!
     
     @IBAction func shortcutPressed(sender: AnyObject) {
-        if let lastWaterLog = getLastWaterLog() {
-            saveWaterLog(lastWaterLog.container!)
-            
+        if let lastWaterLog = WaterLogManager.getLastWaterLog() {
+            WaterLogManager.saveWaterLog(lastWaterLog.container!)
+            self.updateViewForWater()
         }
     }
 
     @IBAction func button1Pressed(sender: AnyObject) {
-        saveWaterLog("sip")
+        WaterLogManager.saveWaterLog("sip")
+        self.updateViewForWater()
     }
     
     @IBAction func button2Pressed(sender: AnyObject) {
-        saveWaterLog("cup")
+        WaterLogManager.saveWaterLog("cup")
+        self.updateViewForWater()
     }
     
     @IBAction func button3Pressed(sender: AnyObject) {
-        saveWaterLog("mug")
+        WaterLogManager.saveWaterLog("mug")
+        self.updateViewForWater()
     }
     
     @IBAction func button4Pressed(sender: AnyObject) {
-        saveWaterLog("bottle")
+        WaterLogManager.saveWaterLog("bottle")
+        self.updateViewForWater()
     }
     
     @IBAction func undoPressed(sender: AnyObject) {
-        undoLastWaterLog()
+        WaterLogManager.undoLastWaterLog()
+        self.updateViewForWater()
     }
 
     override func viewWillAppear(animated: Bool) {
         // Setting up informatinos about water
-        updateSetting()
-        updateWater()
+        self.updateViewForSetting()
+        self.updateViewForWater()
         
         if setting_info.unit == HKUnit(fromString: "mL"){
             sipLabel.text = (setting_info.sipVolume?.doubleValue.toString)! + (setting_info.unit?.description)!
@@ -90,7 +95,6 @@ class ViewController: UIViewController, WCSessionDelegate {
             mugLabel.text = (setting_info.mugVolume?.doubleValue.ml_to_oz.toString)! + (setting_info.unit?.description)!
             bottleLabel.text = (setting_info.bottleVolume?.doubleValue.ml_to_oz.toString)! + (setting_info.unit?.description)!
         }
-        
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -164,68 +168,10 @@ class ViewController: UIViewController, WCSessionDelegate {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
-    // get date in "2015-05-15" form considering timezone.
-    func getDate(date : NSDate) -> String{
-        let dateFormatter = NSDateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        dateFormatter.timeZone = NSTimeZone.defaultTimeZone()
-        let dateString = dateFormatter.stringFromDate(date)
-        return dateString
-    }
-    
-    // fetch water data
-    func fetchWater() -> Double {
-        
-        let fetchRequest = NSFetchRequest(entityName: "WaterLog")
-        
-        var fetchResults = (try? managedObjectContext.executeFetchRequest(fetchRequest)) as? [WaterLog]
-        fetchResults = fetchResults?.reverse()
-        
-        var consumed : Double = 0
-        
-        let today = getDate(NSDate())
 
-        for result in fetchResults! {
-            if (getDate(result.loggedTime!) == today){
-                consumed = consumed + Double(result.amount!)
-            }
-        }
-        
-        return consumed
-    }
-    
-    // store amount of water user consumed in datacore(ml)
-    func saveWaterLog(container : String){
-        let amount = getVolume(container)
-        
-        let unitML: HKUnit = HKUnit(fromString: "mL")
-        let currentUnit: HKUnit = Setting.getUnit()
-        
-        // insert new object into core data framework.
-        let waterLog = NSEntityDescription.insertNewObjectForEntityForName("WaterLog",
-            inManagedObjectContext: managedObjectContext) as! WaterLog
+    // update this view in response to change of setting.
+    func updateViewForSetting() {
 
-        waterLog.amount = amount
-        waterLog.loggedTime = NSDate()
-        waterLog.container = container
-        
-        do {
-            // save the managet object context
-            try managedObjectContext.save()
-            
-            // save HK Sample object for logging drinking water.
-            HealthManager.sharedInstance.requesSavingHKWaterSample(amount)
-        } catch {
-            print("Unresolved error")
-            abort()
-        }
-        // update view in repspons to change of core data objects in WaterLog entity.
-        updateWater()
-    }
-    
-    // update text regards of settings
-    func updateSetting() {
         setting_info = Setting.getSetting()
         if setting_info.unit == HKUnit(fromString: "mL"){
             goal.text = setting_info.goal?.doubleValue.toString
@@ -238,9 +184,10 @@ class ViewController: UIViewController, WCSessionDelegate {
         consumedUnit.text = setting_info.unit?.description
     }
 
-    // update text of consumed water
-    func updateWater() {
-        var consumedWater = fetchWater() as Double
+    // update this view in response to change of water logs
+    func updateViewForWater() {
+
+        var consumedWater = WaterLogManager.getTodayConsumption()
         var goalWater = setting_info.goal?.doubleValue
         if setting_info.unit == HKUnit(fromString: "oz") {
             consumedWater = consumedWater.ml_to_oz
@@ -250,19 +197,18 @@ class ViewController: UIViewController, WCSessionDelegate {
         consumed.text = consumedWater.toString
         
         let progressPercentage = consumedWater / goalWater!
-        let lastWaterLog : WaterLog! = self.getLastWaterLog()
+        let lastWaterLog : WaterLog! = WaterLogManager.getLastWaterLog()
         let waterLeft : Double = goalWater! - consumedWater
-        
 
         // show image of last unit.
         if lastWaterLog != nil {
             let lastContainer = lastWaterLog.container
             var lastContainerVolume : Double = Double()
             if (setting_info.unit == HKUnit(fromString: "mL")){
-                lastContainerVolume = getVolume(lastContainer!)
+                lastContainerVolume = WaterLogManager.getVolume(lastContainer!)
             }
             else if (setting_info.unit == HKUnit(fromString: "oz")){
-                lastContainerVolume = getVolume(lastContainer!).ml_to_oz
+                lastContainerVolume = WaterLogManager.getVolume(lastContainer!).ml_to_oz
             }
             
             let lastUnitImage = UIImage(named: lastContainer! + "_shortcut")
@@ -287,53 +233,6 @@ class ViewController: UIViewController, WCSessionDelegate {
         // cover blue background with white image to show progress status
         waterImageView.image = drawImage(progressPercentage)
     }
-    
-    // Returns core data objects, which is saved in today, in WaterLog entity.
-    func getTodayWaterLogs() -> [WaterLog]? {
-        
-        // today water logs to return.
-        var todayWaterLogs = [WaterLog]()
-        let today = getDate(NSDate())
-        
-        let fetchRequest = NSFetchRequest(entityName: "WaterLog")
-        let fetchResults = (try? managedObjectContext.executeFetchRequest(fetchRequest)) as? [WaterLog]
-        
-        if let results = fetchResults {
-            for result in results {
-                if getDate(result.loggedTime!) == today {
-                    todayWaterLogs.append(result)
-                }
-            }
-        } else {
-            // There's no fetch results, return nil.
-            return nil
-        }
-        
-        return todayWaterLogs
-    }
-    
-    // Delete the last water log
-    func undoLastWaterLog(){
-
-        if let todayWaterLogs = getTodayWaterLogs() {
-            
-            let endIndex = todayWaterLogs.endIndex - 1
-            if endIndex >= 0 {
-                // Delete the last object in WaterLog entity.
-                managedObjectContext.deleteObject(todayWaterLogs[endIndex])
-                // Delete last saved HKSample object meaning drinking water.
-                HealthManager.sharedInstance.requestDeletingLastHKWaterSample()
-            }
-            do {
-                try managedObjectContext.save()
-            } catch {
-                // Do something in response to error condition
-            }
-        }
-        
-        // update view in repspons to change of core data objects in WaterLog entity.
-        updateWater()
-    }
 
     func blurView() {
         let blurView = UIVisualEffectView(effect: UIBlurEffect(style: UIBlurEffectStyle.Dark))
@@ -356,44 +255,10 @@ class ViewController: UIViewController, WCSessionDelegate {
         
         return coverImage
     }
-
-    // Returns the last WaterLog object in core data framework.
-    func getLastWaterLog() -> WaterLog! {
-        
-        let fetchRequest = NSFetchRequest(entityName: "WaterLog")
-        
-        let fetchResults = (try? managedObjectContext.executeFetchRequest(fetchRequest)) as? [WaterLog]
-        if (fetchResults?.count>0){
-            return fetchResults![(fetchResults?.endIndex)!-1]
-        }
-        else {
-            return nil
-        }
-    }
-    
-    // return size(ml) of container.
-    func getVolume(container : String) -> Double {
-        var amount : Double = Double()
-        if container == "sip" {
-            amount = (setting_info.sipVolume?.doubleValue)!
-        }
-        else if container == "cup" {
-            amount = (setting_info.cupVolume?.doubleValue)!
-        }
-            
-        else if container == "mug" {
-            amount = (setting_info.mugVolume?.doubleValue)!
-        }
-        else if container == "bottle" {
-            amount = (setting_info.bottleVolume?.doubleValue)!
-        }
-        return amount
-    }
-    
 }
 
 extension ViewController{
-    
+
     // handle watch
     private func configureWCSession() {
         session?.delegate = self;
@@ -404,7 +269,8 @@ extension ViewController{
         
         //Use this to update the UI instantaneously (otherwise, takes a little while)
         dispatch_async(dispatch_get_main_queue()) {
-            self.saveWaterLog(container)
+            WaterLogManager.saveWaterLog(container)
+            self.updateViewForWater()
         }
     }
     
@@ -414,21 +280,22 @@ extension ViewController{
         switch message["command"] as! String {
             
         case "undo" :
-            undoLastWaterLog()
-            replyHandler(["consumed": fetchWater()])
+            WaterLogManager.undoLastWaterLog()
+            self.updateViewForWater()
+
+            replyHandler(["consumed": WaterLogManager.getTodayConsumption()])
             
         case "fetchStatus" :
-            
             let setting = Setting.getSetting()
             var consumed : Double = Double()
             var goal : Double = Double()
             
             if setting!.unit == HKUnit(fromString: "mL"){
-                consumed = fetchWater()
+                consumed = WaterLogManager.getTodayConsumption()
                 goal = (setting!.goal?.doubleValue)!
             }
             else {
-                consumed = fetchWater().ml_to_oz
+                consumed = WaterLogManager.getTodayConsumption().ml_to_oz
                 goal = (setting!.goal?.doubleValue.ml_to_oz)!
             }
             
@@ -446,6 +313,5 @@ extension ViewController{
         default:
             break
         }
-        
     }
 }
